remove postUserNameAndPassword()
This removes postUserNameAndPassword() from our public API, which performed OAuth using a grant type not officially supported by Mastodon API (issue #11). Unsupported functionality was temporarily moved into the sample apps, pending a rewrite of those at a later time.
