+++ bigbone-rx/src/main/kotlin/social/bigbone/rx/RxApps.kt
@@ -43,27 +43,4 @@ class RxApps(client: MastodonClient) {
             }
         }
     }
-
-    @Deprecated(
-        "This method uses grant_type 'password' which is undocumented in Mastodon API" +
-            " and should not be used in production code.",
-        ReplaceWith("getAccessToken()"),
-        DeprecationLevel.WARNING
-    )
-    fun postUserNameAndPassword(
-        clientId: String,
-        clientSecret: String,
-        scope: Scope,
-        userName: String,
-        password: String
-    ): Single<AccessToken> {
-        return Single.create {
-            try {
-                val accessToken = apps.postUserNameAndPassword(clientId, clientSecret, scope, userName, password)
-                it.onSuccess(accessToken.execute())
-            } catch (throwable: Throwable) {
-                it.onErrorIfNotDisposed(throwable)
-            }
-        }
-    }
 }
